Complete e2e tests based on the docx diagram.
The files testSteps.js and testScreens.js are now obsolete.

The "e2e-process-steps.docx" document shows all screens and links for
this application.  Look at this first.

The "scenarios.js" script now contains tests for all of these in all
combinations. Look at this after looking at the docx. The beast place to
start with this file is at the end where the processes are defined and
run.

I'm still trying to work out how to do some things automatically, but
angularjs-e2e is restricting me to the point that it may just be
impossible to automate some (pretty basic) tests.